feat(cli): move organization app commands to root
