+++ cmd/deletecmd/delete_test.go
