+++ cmd/deletecmd/mock.go
