+++ cmd/deploy/cmd.go
@@ -14,7 +14,7 @@ import (
 var DeployCmd = &cobra.Command{
 	Use:   "deploy [app directory]",
 	Run:   run,
-	Short: "Deploy an app to an organization.",
+	Short: "Deploy an app to an organization",
 	Long: `Deploys an application to an organization on the Numerous platform.
 
 An app's deployment is identified with the <name> and <organization> identifier.
