+++ cmd/legacy/cmd.go
@@ -13,7 +13,7 @@ import (
 
 var LegacyRootCmd = &cobra.Command{
 	Use:   "legacy",
-	Short: "Commands for managing legacy apps on numerous",
+	Short: "Commands for managing legacy apps on Numerous",
 	Args:  args.SubCommandRequired,
 }
