+++ cmd/logs/logs.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"time"
 
-	"numerous.com/cli/cmd/app/appident"
 	"numerous.com/cli/cmd/output"
 	"numerous.com/cli/internal/app"
+	"numerous.com/cli/internal/appident"
 )
 
 type AppService interface {
