+++ internal/app/app_deploy_logs_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"numerous.com/cli/cmd/app/appident"
+	"numerous.com/cli/internal/appident"
 	"numerous.com/cli/internal/test"
 
 	"github.com/stretchr/testify/assert"
